stable

+++ backend/routes/posts.js
@@ -35,6 +35,7 @@ router.post(
       title: req.body.title,
       content: req.body.content,
       imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId,
     });
     post.save().then((created) => {
       res.status(201).json({
@@ -85,9 +86,15 @@ router.put(
       title: req.body.title,
       content: req.body.content,
       imagePath: path,
+      creator: req.userData.userId,
     });
-    Post.updateOne({ _id: req.params.id }, post).then((result) => {
-      res.status(200).json({ message: "Update Successful" });
+    Post.updateOne(
+      { _id: req.params.id, creator: req.userData.userId },
+      post
+    ).then((result) => {
+      if (result.modifiedCount > 0)
+        res.status(200).json({ message: "Update Successful" });
+      else res.status(401).json({ message: "Not authorized" });
     });
   }
 );
@@ -100,9 +107,13 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(() => {
-    res.status(200).json({ message: "Deleted" });
-  });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(
+    (result) => {
+      if (result.deleteCount > 0)
+        res.status(200).json({ message: "Update Successful" });
+      else res.status(401).json({ message: "Not authorized" });
+    }
+  );
 });
 
 module.exports = router;
